DB
Методы для User:  addUser, getUserByLogin, getUserById, getUserById, updateToken
Методы для Message: addMessage, Методы для Bullet: addBullet, DeleteBullet
