+++ Cyborgs.sql
@@ -3,7 +3,7 @@
 -- https://www.phpmyadmin.net/
 --
 -- Хост: 127.0.0.1:3306
--- Время создания: Окт 27 2023 г., 11:23
+-- Время создания: Окт 29 2023 г., 10:00
 -- Версия сервера: 8.0.30
 -- Версия PHP: 8.1.9
 
@@ -24,6 +24,44 @@ SET time_zone = "+00:00";
 -- --------------------------------------------------------
 
 --
+-- Структура таблицы `Bullet`
+--
+
+CREATE TABLE `Bullet` (
+  `id` int NOT NULL,
+  `id_user` varchar(255) COLLATE utf8mb4_general_ci NOT NULL,
+  `x` int NOT NULL,
+  `y` int NOT NULL,
+  `x1` int NOT NULL,
+  `y1` int NOT NULL,
+  `speed` int NOT NULL
+) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_general_ci;
+
+-- --------------------------------------------------------
+
+--
+-- Структура таблицы `Messages`
+--
+
+CREATE TABLE `Messages` (
+  `id` int NOT NULL,
+  `user_id` varchar(255) COLLATE utf8mb4_general_ci NOT NULL,
+  `message` varchar(255) COLLATE utf8mb4_general_ci NOT NULL,
+  `created` varchar(255) COLLATE utf8mb4_general_ci NOT NULL
+) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_general_ci;
+
+--
+-- Дамп данных таблицы `Messages`
+--
+
+INSERT INTO `Messages` (`id`, `user_id`, `message`, `created`) VALUES
+(1, '1', 'иди в жопу', '45'),
+(2, '1', 'иди в жопу', '2023-10-28 20:21:41'),
+(3, '1', 'иди в жопу', '2023-10-28 20:21:43');
+
+-- --------------------------------------------------------
+
+--
 -- Структура таблицы `Users`
 --
 
@@ -37,8 +75,27 @@ CREATE TABLE `Users` (
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_general_ci;
 
 --
+-- Дамп данных таблицы `Users`
+--
+
+INSERT INTO `Users` (`id`, `login`, `password`, `name`, `soname`, `token`) VALUES
+(1, 'вася', '123', 'Игорь', 'Раскольников', '1');
+
+--
 -- Индексы сохранённых таблиц
 --
+
+--
+-- Индексы таблицы `Bullet`
+--
+ALTER TABLE `Bullet`
+  ADD PRIMARY KEY (`id`);
+
+--
+-- Индексы таблицы `Messages`
+--
+ALTER TABLE `Messages`
+  ADD PRIMARY KEY (`id`);
 
 --
 -- Индексы таблицы `Users`
@@ -51,10 +108,22 @@ ALTER TABLE `Users`
 --
 
 --
+-- AUTO_INCREMENT для таблицы `Bullet`
+--
+ALTER TABLE `Bullet`
+  MODIFY `id` int NOT NULL AUTO_INCREMENT, AUTO_INCREMENT=9;
+
+--
+-- AUTO_INCREMENT для таблицы `Messages`
+--
+ALTER TABLE `Messages`
+  MODIFY `id` int NOT NULL AUTO_INCREMENT, AUTO_INCREMENT=4;
+
+--
 -- AUTO_INCREMENT для таблицы `Users`
 --
 ALTER TABLE `Users`
-  MODIFY `id` int NOT NULL AUTO_INCREMENT;
+  MODIFY `id` int NOT NULL AUTO_INCREMENT, AUTO_INCREMENT=2;
 COMMIT;
 
 /*!40101 SET CHARACTER_SET_CLIENT=@OLD_CHARACTER_SET_CLIENT */;
