--- server/application/modules/DB/config.php
@@ -1,10 +1,0 @@
-<?php
-
-return [
-    'host'=>'localhost',
-    'db_name'=>'Cyborgs',
-    'username'=>'root',
-    'password'=>'',
-    'charset'=>'utf8',
-    
-];
